--- src/main/java/com/eedo/ZipTe/domain/apt/entity/ResidenceType.java
@@ -1,7 +1,0 @@
-package com.eedo.ZipTe.domain.apt.entity;
-
-public enum ResidenceType {
-
-    APARTMENT,VILLA,HOUSE
-
-}
