--- src/main/java/com/eedo/ZipTe/domain/apt/entity/bulletin/Review.java
@@ -1,42 +1,0 @@
-package com.eedo.ZipTe.domain.apt.entity.bulletin;
-
-import com.eedo.ZipTe.domain.apt.entity.Apt;
-import com.eedo.ZipTe.domain.member.entity.member.Member;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-
-@Entity
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Review {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name ="member_id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "apt_id")
-    private Apt apt;
-
-    private String content;
-
-    private int rating;
-
-    @Temporal(TemporalType.TIMESTAMP)
-    private Date createdDate;
-
-    @Builder.Default
-    @OneToMany(mappedBy = "review", cascade = CascadeType.ALL)
-    private List<Comment> commentList = new ArrayList<>();
-}
