--- src/main/java/com/eedo/ZipTe/domain/apt/repo/AptRepository.java
@@ -1,8 +1,0 @@
-package com.eedo.ZipTe.domain.apt.repo;
-
-import com.eedo.ZipTe.domain.apt.entity.Apt;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AptRepository extends JpaRepository<Apt, Long> {
-
-}
