--- src/main/java/com/eedo/ZipTe/domain/apt/service/AptService.java
@@ -1,16 +1,0 @@
-package com.eedo.ZipTe.domain.apt.service;
-
-import com.eedo.ZipTe.domain.apt.repo.AptRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class AptService {
-
-    private final AptRepository aptRepository;
-
-
-
-
-}
