+++ src/main/java/com/eedo/ZipTe/domain/entity/Address.java
@@ -1,0 +1,22 @@
+package com.eedo.ZipTe.domain.entity;
+
+import jakarta.persistence.Embeddable;
+import jakarta.persistence.Enumerated;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Embeddable
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class Address {
+
+    private String city;
+    private String roadName;
+    private String detail;
+    private int zipcode;
+
+}
