+++ src/main/java/com/eedo/ZipTe/domain/entity/apt/ResidenceType.java
@@ -1,0 +1,7 @@
+package com.eedo.ZipTe.domain.entity.apt;
+
+public enum ResidenceType {
+
+    APARTMENT,VILLA,HOUSE
+
+}
