+++ src/main/java/com/eedo/ZipTe/domain/entity/bulletin/Comment.java
@@ -1,0 +1,32 @@
+package com.eedo.ZipTe.domain.entity.bulletin;
+
+import com.eedo.ZipTe.domain.entity.member.Member;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Comment {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "review_id")
+    private Review review;
+
+    private String content;
+
+    private LocalDateTime createdDate;
+}
