+++ src/main/java/com/eedo/ZipTe/domain/entity/item/NotEnoughStockException.java
@@ -1,0 +1,16 @@
+package com.eedo.ZipTe.domain.entity.item;
+
+public class NotEnoughStockException extends RuntimeException {
+    public NotEnoughStockException() {
+    }
+    public NotEnoughStockException(String message) {
+        super(message);
+    }
+    public NotEnoughStockException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public NotEnoughStockException(Throwable cause) {
+        super(cause);
+    }
+}
