+++ src/main/java/com/eedo/ZipTe/domain/entity/member/MemberRole.java
@@ -1,0 +1,6 @@
+package com.eedo.ZipTe.domain.entity.member;
+
+public enum MemberRole {
+    USER_GREEN,USER_VIP,MANAGER,ADMIN,
+
+}
