+++ src/main/java/com/eedo/ZipTe/domain/entity/order/DeliveryStatus.java
@@ -1,0 +1,6 @@
+package com.eedo.ZipTe.domain.entity.order;
+
+public enum DeliveryStatus {
+
+    READY,COMP
+}
