+++ src/main/java/com/eedo/ZipTe/domain/entity/order/OrderStatus.java
@@ -1,0 +1,6 @@
+package com.eedo.ZipTe.domain.entity.order;
+
+public enum OrderStatus {
+
+    ORDER , CANCEL
+}
