+++ src/main/java/com/eedo/ZipTe/domain/entity/owner/OwnerState.java
@@ -1,0 +1,6 @@
+package com.eedo.ZipTe.domain.entity.owner;
+
+public enum OwnerState {
+
+    매매,전세,월세
+}
