--- src/main/java/com/eedo/ZipTe/domain/item/entity/Item.java
@@ -1,41 +1,0 @@
-package com.eedo.ZipTe.domain.item.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class Item {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "item_id")
-    private Long id;
-
-    private String name;
-
-    private int price;
-
-    private int stockQuantity;
-
-
-    // 비즈니스 로직
-    public void addStock(int quantity) {
-        this.stockQuantity += quantity;
-    }
-
-    public void removeStock(int quantity) {
-        int restStock = this.stockQuantity - quantity;
-        if (restStock < 0) {
-            throw new NotEnoughStockException("need More stock");
-        }
-        this.stockQuantity = restStock;
-    }
-
-}
