--- src/main/java/com/eedo/ZipTe/domain/item/entity/NotEnoughStockException.java
@@ -1,16 +1,0 @@
-package com.eedo.ZipTe.domain.item.entity;
-
-public class NotEnoughStockException extends RuntimeException {
-    public NotEnoughStockException() {
-    }
-    public NotEnoughStockException(String message) {
-        super(message);
-    }
-    public NotEnoughStockException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public NotEnoughStockException(Throwable cause) {
-        super(cause);
-    }
-}
