--- src/main/java/com/eedo/ZipTe/domain/member/dto/MemberDTO.java
@@ -1,56 +1,0 @@
-package com.eedo.ZipTe.domain.member.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-//@Data
-//@Builder
-public class MemberDTO extends User {
-
-    private String email;
-    private String pw;
-    private String nickName;
-
-    private boolean social;
-
-    private List<String> roleNames = new ArrayList<>();
-
-    public MemberDTO(String email, String pw, String nickname, boolean social, List<String> roleNames) {
-
-        super(
-                email,
-                pw,
-                roleNames.stream().map(str -> new SimpleGrantedAuthority("ROLE_" + str)).collect(Collectors.toList()));
-
-        this.email = email;
-        this.pw = pw;
-        this.nickName = nickname;
-        this.social = social;
-        this.roleNames = roleNames;
-    }
-
-    // JWT라는 문자를 통해 통신을 한다.
-    // DTO를 바꿔서 JWT문자로 바꾸는 기능을 추가
-    public Map<String, Object> getClaims() {
-        Map<String, Object> dataMap = new HashMap<>();
-
-        dataMap.put("email", email);
-        dataMap.put("pw", pw);
-        dataMap.put("nickname", nickName);
-        dataMap.put("social", social);
-        dataMap.put("roleNames", roleNames);
-
-        return dataMap;
-    }
-
-}
