--- src/main/java/com/eedo/ZipTe/domain/member/entity/member/MemberRole.java
@@ -1,8 +1,0 @@
-package com.eedo.ZipTe.domain.member.entity.member;
-
-import lombok.Getter;
-
-public enum MemberRole {
-    USER_GREEN,USER_VIP,MANAGER,ADMIN,
-
-}
