--- src/main/java/com/eedo/ZipTe/domain/member/entity/owner/OwnerState.java
@@ -1,6 +1,0 @@
-package com.eedo.ZipTe.domain.member.entity.owner;
-
-public enum OwnerState {
-
-    매매,전세,월세
-}
