--- src/main/java/com/eedo/ZipTe/domain/member/repo/MemberRepository.java
@@ -1,16 +1,0 @@
-package com.eedo.ZipTe.domain.member.repo;
-
-
-import com.eedo.ZipTe.domain.member.entity.member.Member;
-import org.springframework.data.jpa.repository.EntityGraph;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, String> {
-    @EntityGraph(attributePaths = {"memberRoleList"})
-    @Query("select m from Member m where m.email =:email")
-    Optional<Member> getWithRoles(@Param("email") String email);
-}
