--- src/main/java/com/eedo/ZipTe/domain/order/entity/Delivery.java
@@ -1,35 +1,0 @@
-package com.eedo.ZipTe.domain.order.entity;
-
-import com.eedo.ZipTe.domain.member.entity.Address;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class Delivery {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "delivery_id")
-    private long id;
-
-    @Embedded
-    private Address address;
-
-    @OneToOne(mappedBy = "delivery",fetch = FetchType.LAZY)
-    private Order order;
-
-    @Enumerated(EnumType.ORDINAL)
-    private DeliveryStatus deliveryStatus;
-
-//    비즈니스 모델
-    public void addOrder(Order order){
-        this.order = order;
-    }
-}
