--- src/main/java/com/eedo/ZipTe/domain/order/entity/DeliveryStatus.java
@@ -1,6 +1,0 @@
-package com.eedo.ZipTe.domain.order.entity;
-
-public enum DeliveryStatus {
-
-    READY,COMP
-}
