--- src/main/java/com/eedo/ZipTe/domain/order/entity/OrderStatus.java
@@ -1,8 +1,0 @@
-package com.eedo.ZipTe.domain.order.entity;
-
-import jakarta.persistence.Embeddable;
-
-public enum OrderStatus {
-
-    ORDER , CANCEL
-}
