+++ src/main/java/com/eedo/ZipTe/domain/repository/AptRepository.java
@@ -1,0 +1,8 @@
+package com.eedo.ZipTe.domain.repository;
+
+import com.eedo.ZipTe.domain.entity.apt.Apt;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AptRepository extends JpaRepository<Apt, Long> {
+
+}
