+++ src/main/java/com/eedo/ZipTe/domain/service/AptService.java
@@ -1,0 +1,16 @@
+package com.eedo.ZipTe.domain.service;
+
+import com.eedo.ZipTe.domain.repository.AptRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class AptService {
+
+    private final AptRepository aptRepository;
+
+
+
+
+}
